sportsight_mvp/requirements.txt
ultralytics
opencv-python
ffmpeg-python
python-dotenv
requests
transformers
torch
openai 